@@ -1,29 +1,29 @@
 import React, {useEffect} from "react";
 
 const SpeechAudio = ({word}) => {
-
-
-  useEffect( () => {
+  useEffect(() => {
     const createSpeech = () => {
-      let languages = ['en-GB','en-US','en-AU','es-ES','cs-CZ','it-IT','pl-PL','ru-RU','fr-FR'];
+      let languages = ["en-GB", "en-US", "en-AU"];
       let speech = new SpeechSynthesisUtterance();
 
       let languageIndex = Math.floor(languages.length * Math.random());
       speech.text = word;
-      speech.lang = languages[0];
+      speech.lang = languages[languageIndex];
       speech.pitch = 2 * Math.random();
 
       return speech;
-    }
+    };
 
-    if('speechSynthesis' in window){
+    if ("speechSynthesis" in window) {
       window.speechSynthesis.speak(createSpeech());
     } else {
-      console.warn('The current browser does not support the speechSynthesis API.')
+      console.warn(
+        "The current browser does not support the speechSynthesis API."
+      );
     }
-  },[]);
+  }, []);
 
-	return null;
-}
+  return null;
+};
 
-export default SpeechAudio
+export default SpeechAudio;
